making changes in order to remodel the authentication structure and to fix the suspected security issue
