+++ shibboleth/backends.py
@@ -1,6 +1,9 @@
 from django.db import connection
 from django.contrib.auth.models import User, Permission
 from django.contrib.auth.backends import RemoteUserBackend
+from shibboleth.app_settings import SHIB_ATTRIBUTE_MAP
+import re
+
 
 class ShibbolethRemoteUserBackend(RemoteUserBackend):
     """
@@ -17,7 +20,7 @@ class ShibbolethRemoteUserBackend(RemoteUserBackend):
     # Create a User object if not already in the database?
     create_unknown_user = True
 
-    def authenticate(self, remote_user, shib_meta):
+    def authenticate(self, remote_user, META_HEADERS,session):
         """
         The username passed as ``remote_user`` is considered trusted.  This
         method simply returns the ``User`` object with the given username,
@@ -29,7 +32,19 @@ class ShibbolethRemoteUserBackend(RemoteUserBackend):
         if not remote_user:
             return
         user = None
-        username = self.clean_username(remote_user)
+        username = remote_user
+        # the fact that username is not checked against the shibboleth-ID is a security hole here.
+
+
+        # Make sure we have all required Shiboleth elements before proceeding.
+        shib_meta, error = self.parse_attributes(META_HEADERS)
+        # Add parsed attributes to the session. # Why is this necessary??
+        session['shib'] = shib_meta
+
+        if error:
+            raise ShibbolethValidationError("All required Shibboleth elements"
+                                            " not found.  %s" % shib_meta)
+
         shib_user_params = dict([(k, shib_meta[k]) for k in User._meta.get_all_field_names() if k in shib_meta])
         # Note that this could be accomplished in one try-except clause, but
         # instead we use get_or_create when creating unknown users since it has
@@ -44,3 +59,33 @@ class ShibbolethRemoteUserBackend(RemoteUserBackend):
             except User.DoesNotExist:
                 pass
         return user
+
+    def clean_username(self,value):
+        # find relevant substring of shibboleth attribute
+        regex = re.compile("de/shibboleth\!(.*)=")
+        value = regex.findall(value)[-1]
+        # remove special characters
+        value = ''.join(e for e in value if e.isalnum())
+        return value
+
+    def parse_attributes(self, meta):
+        """
+        Parse the incoming Shibboleth attributes.
+        From: https://github.com/russell/django-shibboleth/blob/master/django_shibboleth/utils.py
+        Pull the mapped attributes from the apache headers.
+        """
+        shib_attrs = {}
+        error = False
+        for header, attr in SHIB_ATTRIBUTE_MAP.items():
+            required, name = attr
+            value = meta.get(header, None)
+            if name == "username":
+                value = self.clean_username(value)
+            shib_attrs[name] = value
+            if not value or value == '':
+                if required:
+                    error = True
+        return shib_attrs, error
+
+class ShibbolethValidationError(Exception):
+    pass
